moved files, added class Parser
