+++ src/main/java/homework/BusTicket.java
@@ -1,4 +1,4 @@
-package homework4;
+package homework;
 
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
