+++ src/main/java/homework/Parser.java
@@ -1,18 +1,17 @@
-package homework4;
+package homework;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 
-
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
-
-public class JsonParser {
-    public static void main(String[] args) {
-        List<BusTicket> tickets;
+public class Parser {
+    public List<BusTicket> parseJSON() {
+        List<BusTicket> tickets = new ArrayList<>();
         String filePath = "tickets.json";
         try {
             ObjectMapper mapper = new ObjectMapper();
@@ -21,22 +20,9 @@ public class JsonParser {
                     new TypeReference<List<BusTicket>>() {
                     });
             System.out.println(tickets);
-
-            TicketValidator validator = new TicketValidator();
-            int validTickets = 0;
-
-            for (BusTicket ticket : tickets) {
-                List<Integer> violationCount = validator.checkTicket(ticket);
-                if(violationCount.isEmpty()) {
-                    validTickets++;
-                }
-            }
-            System.out.println("Total: " + tickets.size());
-            System.out.println("Valid: " + validTickets);
-            validator.printViolation();
-
         } catch (IOException e) {
             e.printStackTrace();
         }
+        return tickets;
     }
 }
