+++ src/main/java/homework/TicketService.java
@@ -1,8 +1,8 @@
-package homework2;
+package homework;
 
-import homework2.user.Admin;
-import homework2.user.Client;
-import homework2.user.User;
+import homework.user.Admin;
+import homework.user.Client;
+import homework.user.User;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -75,5 +75,25 @@ public class TicketService {
             }
         }
         return ticketsBySector;
+    }
+    public void validateBusTickets() {
+        List<BusTicket> tickets;
+
+        Parser parser = new Parser();
+        tickets = parser.parseJSON();
+
+        TicketValidator validator = new TicketValidator();
+        int validTickets = 0;
+
+        for (BusTicket ticket : tickets) {
+            List<Integer> violationCount = validator.checkTicket(ticket);
+            if (violationCount.isEmpty()) {
+                validTickets++;
+            }
+        }
+        System.out.println("Total: " + tickets.size());
+        System.out.println("Valid: " + validTickets);
+        validator.printViolation();
+
     }
 }
